refactor(sdk): split canon ABI variant tests
The canon ABI integration module now hosts shared dynamic project helpers used by multiple component-model test groups.

Move the enum variant cases and their WIT assertions into a dedicated variant submodule beside the option tests. This keeps the parent module focused on account and note project construction while preserving the existing variant coverage.
